+++ projects/map-maplibre/src/lib/map-maplibre.service.ts
@@ -209,9 +209,11 @@ export class MapMaplibreService {
   }
 
   private updateLayerParamsAndSource(map: glMap, mllayer: StyleLayer, layer: ukisLayer) {
-    if (layer.type === 'wms' || layer.type === 'wmts' || layer.type === 'tms' || layer.type === 'wfs' || layer.type === 'geojson') {
+    if (['wms', 'wmts', 'tms', 'wfs', 'geojson', 'custom'].includes(layer.type)) {
       const oldSource = map.getSource(mllayer.source);
-      updateSource(map, layer, oldSource);
+      if (oldSource) {
+        updateSource(map, layer, oldSource);
+      }
     }
   }