+++ CHANGELOG.md
@@ -32,6 +32,7 @@
  - Allow the usage of different coordinate reference systems in WMTS requests. Use `WmtsLayer.params.matrixSetOptions.matrixSet`. and `WmtsLayer.params.matrixSetOptions.tileMatrixPrefix`
  - Add more types and fix type errors
  - Fix KMLSource `setData` add `waitForCompletion`
+ - Update `custom` source in `updateLayerParamsAndSource` to also update custom layers
 
 
 # [16.0.0](https://github.com/dlr-eoc/ukis-frontend-libraries/tree/v16.0.0) (2026-05-29) (update dependencies, new library ngx-ukis-ui-clarity and ngx-ukis-utilities, replace core-ui)